Fix Linux host mouse button mappings.
For global mappings (xmodmap), we try to adapt to them (i.e. try to use a reverse map), so that we don't mess with the user's environment in the IT2Me scenario.

For device mappings (xinput), I think we can assume that it's safe to assign a sane (identity) mapping to XTEST. As far as I can tell, the only reason it could be different is if it's changed by mistake by GNOME2's settings daemon.

Right now it fixes the mappings on client connect (meaning that if the user changes mappings while a connection is running, it'll get inverted until). We can guard against this by listening to the change events, but that requires:
  (a) creating another thread to listen for events
  (b) hooking up one of the existing event handlers (video_frame_capturer_linux) to event_executor_linux
  or (c) refactoring event handlers

Looking at Lambros's change for the clipboard, it seems that there is some event handling thread that can be shared (either the thread itself, or at least the code to do that via file descriptor watcher), so I'll wait for his input.

BUG=126490
TEST=Manual ('xmodmap -e "pointer = 3 2 1"', 'xinput set-button-map <xtest_device_id> 3 2 1' to emulate both situations, as my linux machine runs GNOME3)


Review URL: https://chromiumcodereview.appspot.com/10905221

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@156594 0039d316-1c4b-4281-b951-d872f2087c98