Instant-extended: Send hash state.
When the active tab is pointed at the same host and path as the Instant
URL, propagate its hash state to the overlay. This is an attempt to keep
search modes around.

I am not sure this is going to cut it due to these problems:
- This doesn't work for the first Google Images mode search following a
  switch into Images mode, because it doesn't yet have hash state, just
  ordinary query parameters.
- There is some kind of race where if you refine from [a] to [b]
  quickly in Google, we send the hash state but it is not applied to
  [b]; the state is applied reliably if [a] works.

BUG=135294

Review URL: https://chromiumcodereview.appspot.com/10917145

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@157386 0039d316-1c4b-4281-b951-d872f2087c98