Simple mac curtain mode fix.
This just moves the curtain mode policy last (so there'll be no restarts after it's applied). Looking through the code, this should be enough to ensure that the policy stays applied.

This also fixes the threading on OnDisconnectRequested - it uses the host_ pointer, so it should run from the network thread and check the pointer, to make sure the host_ pointer doesn't go away from underneath it.


BUG=150707

TEST=Ad Hoc (add policy to macbook, remote into it, verify macbook session is suspended. Also, test that the UI disconnect button still works)


Review URL: https://chromiumcodereview.appspot.com/10957062

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@158454 0039d316-1c4b-4281-b951-d872f2087c98