Change a result code ENUM value to help diagnose unexpected histogram data
We've gotten histogram data showing the integer value of 24, which happens to
be the same as RESULT_CODE_INSTALL_FROM_WEBSTORE_ERROR, as the exit code for
some renderer crashes. This seems impossible since that result code should only
be used for browser process exit very early in startup, when a special
 --install-from-webstore experimental command line flag is present.

What seems more likely is that some other exit values are getting conflated with
this one somewhere in the reporting pipeline, perhaps because it comes last in
the list. To verify this hypothesis, I'm changing the name and value of
RESULT_CODE_INSTALL_FROM_WEBSTORE_ERROR, adding a dummy value in its place, as
well as following it with another dummy value. 

BUG=152285

Review URL: https://chromiumcodereview.appspot.com/10987038

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@158769 0039d316-1c4b-4281-b951-d872f2087c98