Fix up instance registration with the browser.
This code was crashing and was disabled. Previously it would try to send the message through the render view passed to the constructor. But since the dispatcher wrapper outlives any particular RenderView, this crashes.

This new code goes through the RendererPpapiHost to get the render view associated with the current instance that's being created or destroyed.

BUG=150654

Review URL: https://codereview.chromium.org/10949008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@158889 0039d316-1c4b-4281-b951-d872f2087c98