[Mac]: Fix "Look Up in Dictionary" and "Start Speaking" not working for PDFs.
This was broken by http://crrev.com/150388, which refactored the dictionary
and speech code to go through RWHVM. Unfortunately, this broke things
because getting an attributed string from the PDF is not implemented, and
because the PDF plugin did not update RWHVM with the current text selection.

This change works around these issues by updating RWHVM's text selection
when the context menu is brought up and taking the legacy "Bring up
Dictionary.app" path (which had been removed by http://crrev.com/150388)
in the case when an attributed string is not available.

BUG=151641
TEST=Go to http://mysite.science.uottawa.ca/rsmith43/Zombies.pdf. Highlight
"Zombies" and right click on it. Choose "Look Up in Dictionary". The dictionary
app should be opened showing the definition for Zombies. Right click again
and choose "Speech" -> "Start Speaking". "Zombies" should be spoken.
On Lion+, go to a regular HTML website and right click on a word and choose
"Look Up in Dictionary". The Dictionary _panel_ (not app), should still be
brought up.

Review URL: https://chromiumcodereview.appspot.com/10993014

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@158988 0039d316-1c4b-4281-b951-d872f2087c98