Add ActionBoxController, a controller class for the action box scoped to the lifetime of the box itself, as opposed to ActionBoxMenuModel which is scoped to the lifetime of the menu. Move business logic out of ActionBoxMenuModel and into ActionBoxController.
The reason we need this is because some extension logic needs to be scoped to
the action box -- for example, we will need the commands API to affect the
action box when there isn't a menu open. We may also want to do something like
open up the action box and point at the extension when one is installed.
This logic could be implemented in UI code but that would be icky.


BUG=125307

Review URL: https://chromiumcodereview.appspot.com/10979014

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@159201 0039d316-1c4b-4281-b951-d872f2087c98