Add web page replay to chrome_remote_control.
This adds an "archive_path" argument to the PageSet which specifies the
archive to record to or replay from. Specifying it there will allow for
sharding when page sets are broken into smaller subsets.

By default the benchmark still uses the live sites. If --record is passed, the
benchmark is run and web content is recorded to the archive. If an archive
exists, web content only comes from the archive.

BUG=None
TEST=Manual testing on linux

Review URL: https://codereview.chromium.org/10965027

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@159714 0039d316-1c4b-4281-b951-d872f2087c98