Added sync calls to resources.
The resource (on the plugin side) provides an outgoing and a reply IPC message and the reply message will be set up synchronously. The nice thing about this change is that when writing new ResourceHosts, there is no difference between handling a sync message and a non-sync message. So, for example, you could have a single implementation of some function on the host side and the plugin could choose to call that implementation both synchronously and asynchronously. For example, for the function GetDefaultPrintSettings(), the plugin could call both:
CallBrowser(PpapiHostMsg_Printing_GetDefaultPrintSettings());
AND
CallBrowserSync(PpapiHostMsg_Printing_GetDefaultPrintSettings(), &reply);

with only a single implementation of GetDefaultPrintSettings on the host side.

This change also supports both synchronously and asynchronous handling sync messages on the host side (through IPC_MESSAGE_HANDLER_DELAY_REPLY).


BUG=none


Review URL: https://chromiumcodereview.appspot.com/10989042

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@159820 0039d316-1c4b-4281-b951-d872f2087c98