Browser Plugin: Fixed browser process crash on embedder reload.
The browser process crashed because a CHECK that verifies that an existing browser_plugin_embedder_ does not yet exist fails.

This happens because CreateGuest should only be handled in WebContentsImpl on the first call to CreateGuest. On all subsequent calls, CreateGuest should be handled in BrowserPluginEmbedderHelper.

This patch makes this change.

I've also cleaned up some dead code.

I've also written a test that verifies that reloading the embedder
doesn't cause the browser plugin to misbehave.


BUG=153561
Test=BrowserPluginHostTest.ReloadEmbedder


Review URL: https://chromiumcodereview.appspot.com/11046002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@159854 0039d316-1c4b-4281-b951-d872f2087c98