<browser> Disable force_compositing on guests. We don't yest have compositing path implemented for browser plugin, this will allow us to switch to new path and delete old path once things are stable.
About making the change in web_contents: we have to do it after chrome/browser/content_browser_client overrides the flag (OverrideWebkitPrefs()), not totally sure if this is the right place to hook it up. The comment for OverrideWebkitPrefs says: "The content layer will add its own settings, and then it's up to the embedder to update it if it wants." --> does this mean it is OK to override it in web_contents?

BUG=152884
TEST=With force compositing flag turned on, noticed that pages with guests now render. Also tested that apps would work.


Review URL: https://chromiumcodereview.appspot.com/11035008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@159933 0039d316-1c4b-4281-b951-d872f2087c98