This CL supersedes previous ConstrainedWindow Views CLs and contains much of the work required to update ConstrainedWindowViews to meet the mock specified at: http://www.corp.google.com/~kenmoore/mocks/chromeos/Misc_2012/Dialogs/Markup2/dialogs_markup2g.html
I believe that all the features of the mock are implemented by this CL,
except:
 - dialog shadow
 - dialog animations
 - button shadows
 - some aspects of button text layout

Note that this CL attempts to fully address *only* the repost form
warning dialog.  The intention is to complete the remaining work
(features and dialogs) in future CLs.  The ConstrainedWindow changes are
behind a flag to allow for incremental progress.

The first patch set contains changes to position the constrained window
so that it overlaps the browser chrome.

BUG=140517, 140537, 140539, 140554, 140562


Review URL: https://chromiumcodereview.appspot.com/10933085

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@159994 0039d316-1c4b-4281-b951-d872f2087c98