Android: Add list of files with licenses not wanted in webview.
The script which merges Chromium code into the Android tree needs to
know which directories to remove for licensing reasons; it's much easier
if this is just a static list in the tree which can subsequently be
validated with the webview_licences tool. The list should change very
rarely.

R=mnaganov@chromium.org
BUG=

Review URL: https://codereview.chromium.org/11040056

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@160369 0039d316-1c4b-4281-b951-d872f2087c98