Divide gesture events by dpi scale if IsDIPEnabled().
In the future, we're planning to flip the value of IsDIPEnabled() to
true to match other platforms.  One of the behaviors implied by this is
for gesture events to be divided by the dpi scale in the browser process.

Also, this patch removes the dependency on
WebInputEventFactory::gestureEvent() since it provides no value-added on
Android and has been making it harder to make changes to this code.

BUG=152505


Review URL: https://chromiumcodereview.appspot.com/11099007

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@160940 0039d316-1c4b-4281-b951-d872f2087c98