Fix a crash on Google Drive operations cancelled during reauthenticaiton.
This patch introduces a weak pointer to safely notify to the authenticater
that an operations is terminated by the user.

The operations are terminated for three reasons: (1) user cancel, (2)
succesful or failure) completion, and (3) an authentification failure,
later separated from (2).
 - (1) immediately shuts down (2) by URLFetcher::Cancel().
 - (2) and (3) immediately unrolls the operation so (1) won't happen.
 - The steps (2) and (3) are never simultaneously run, so no problem.
So, the only problematic case was the event (1) followed by (3), in
that case a deleted operation object was used.

BUG=153003


Review URL: https://chromiumcodereview.appspot.com/11032031

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@161036 0039d316-1c4b-4281-b951-d872f2087c98