Fix ppapi TwoWayTest so that it constructs a unique IPC::ChannelHandle.
When multiple TwoWayTests (run as ppapi_unittests) were run concurrently,
several IPC ChannelHandles may have been created with the same name. This would
cause the creation of the channel to fail (because the channel handle has to be
unique). This was causing flaky tests on the trybots. This uses the process ID
in the channel handle name so that they are guaranteed to be unique for the set
of running processes.

This also re-enables a test which I disabled under the suspiscion that it was
related to this failure (but it isn't).

BUG=155175


Review URL: https://chromiumcodereview.appspot.com/11196002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@162458 0039d316-1c4b-4281-b951-d872f2087c98