Tell profile sync svc to use all datatypes when the user chooses the default sync options with one click sign in.
BUG=154940
TEST=Create a new profile.  Don't use sync promo.  Sign in to gmail and use
the one-click sign in feature.  In the bubble that shows up pointing to the
wrench menu, make sure to use the default option and not the advanced option.
Stop and restart chrome.  Go to the settings page and click the "Advanced
sync settings..." button.  Make sure all data types are selected.


Review URL: https://chromiumcodereview.appspot.com/11187030

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@162479 0039d316-1c4b-4281-b951-d872f2087c98