Browser Plugin: Fix Browser Tag crash in Chrome Apps
When we close the Chrome Apps window, RenderViewImpl is destroyed before
the Render tree (and hence BrowserPlugin) is destroyed. This causes
the BrowserPlugin to access an invalid pointer.

This patch addresses this issue by caching the RenderViewImpl's routing ID which
is needed to inform the browser process that the plugin has been destroyed.

This patch also makes similar lifetime issues easier to track by having BrowserPlugin
hold a weak pointer to RenderViewImpl, rather than a raw pointer.

BUG=156198


Review URL: https://chromiumcodereview.appspot.com/11194034

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@162554 0039d316-1c4b-4281-b951-d872f2087c98