[sync] Isolate sync_integration_tests and its dependencies (mac)
In order to run a chrome test suite on a buildbot without checking out
the entire chromium source tree, we need to copy over the transtitive
closure of the test suite and all its dependencies. Doing so manually is
tedious, error prone, and brittle.

This patch updates the .isolate file for sync_integration_tests and modifies
its corresponding test target in chrome_tests.gypi. This will enable us to
copy sync_integration_tests to other tester buildbots in order to run
them as part of cross platform integration tests. We plan on using this
on the bling buildbots for sync.

TBR=thakis@chromium.org
BUG=155531, b/6942067

Review URL: https://codereview.chromium.org/11195035

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@162738 0039d316-1c4b-4281-b951-d872f2087c98