Upstream Blocked Popup Content Settings changes
Since blocked popups rely on an info bar instead of a persistent button, we have to handle
the case where the info bar might have been dismissed and make sure to show it again.  So
every time we block a popup we should notify the WebContents so the front end has a chance
to handle the event and make sure an info bar is showing.

BUG=http://crbug.com/155685


Review URL: https://chromiumcodereview.appspot.com/11238008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@163358 0039d316-1c4b-4281-b951-d872f2087c98